Dev/bryon spectrogram update (#102)
* wip

* requirements fix

* update spectro

* compressed spectrogram model

* linting

* change spectrogram to utilzie URLs instead of base64

* update viewer defaults

* setup update, pip upgrade, merged testing

---------
