+++ bats_ai/core/admin/__init__.py
@@ -1,4 +1,5 @@
 from .annotations import AnnotationsAdmin
+from .compressed_spectrogram import CompressedSpectrogramAdmin
 from .grts_cells import GRTSCellsAdmin
 from .image import ImageAdmin
 from .recording import RecordingAdmin
@@ -14,4 +15,5 @@ __all__ = [
     'TemporalAnnotationsAdmin',
     'SpeciesAdmin',
     'GRTSCellsAdmin',
+    'CompressedSpectrogramAdmin',
 ]
