+++ bats_ai/core/models/__init__.py
@@ -1,4 +1,5 @@
 from .annotations import Annotations
+from .compressed_spectrogram import CompressedSpectrogram
 from .grts_cells import GRTSCells
 from .image import Image
 from .recording import Recording, colormap
@@ -17,4 +18,5 @@ __all__ = [
     'TemporalAnnotations',
     'GRTSCells',
     'colormap',
+    'CompressedSpectrogram',
 ]
