+++ bats_ai/core/models/recording.py
@@ -2,6 +2,7 @@ import logging
 
 from django.contrib.auth.models import User
 from django.contrib.gis.db import models
+from django.dispatch import receiver
 from django_extensions.db.models import TimeStampedModel
 
 from .species import Species
@@ -67,17 +68,17 @@ class Recording(TimeStampedModel, models.Model):
 
     @property
     def spectrogram(self):
-        from bats_ai.core.models import Spectrogram
+        pass
 
         spectrograms = self.spectrograms
-
-        if len(spectrograms) == 0:
-            Spectrogram.generate(self, colormap=COLORMAP)
-
-            spectrograms = self.spectrograms
-            assert len(spectrograms) == 1
 
         assert len(spectrograms) >= 1
         spectrogram = spectrograms[0]  # most recently created
 
         return spectrogram
+
+
+@receiver(models.signals.pre_delete, sender=Recording)
+def delete_content(sender, instance, **kwargs):
+    if instance.audio_file:
+        instance.audio_file.delete(save=False)
