+++ bats_ai/core/models/spectrogram.py
@@ -6,16 +6,17 @@ import math
 from PIL import Image
 import cv2
 from django.core.files import File
+from django.core.files.storage import default_storage
 from django.db import models
 from django.db.models.fields.files import FieldFile
+from django.dispatch import receiver
 from django_extensions.db.models import TimeStampedModel
 import librosa
 import matplotlib.pyplot as plt
 import numpy as np
-import scipy
 import tqdm
 
-from bats_ai.core.models import Annotations, Recording
+from .recording import Recording
 
 logger = logging.getLogger(__name__)
 
@@ -114,7 +115,7 @@ class Spectrogram(TimeStampedModel, models.Model):
         vmin = window.min()
         vmax = window.max()
 
-        chunksize = int(5e3)
+        chunksize = int(2e3)
         arange = np.arange(chunksize, window.shape[1], chunksize)
         chunks = np.array_split(window, arange, axis=1)
 
@@ -218,7 +219,7 @@ class Spectrogram(TimeStampedModel, models.Model):
         img.save(buf, format='JPEG', quality=80)
         buf.seek(0)
 
-        name = 'spectrogram.jpg'
+        name = f'{recording.pk}_{colormap}_spectrogram.jpg'
         image_file = File(buf, name=name)
 
         spectrogram = cls(
@@ -232,148 +233,7 @@ class Spectrogram(TimeStampedModel, models.Model):
             colormap=colormap,
         )
         spectrogram.save()
-
-    @property
-    def compressed(self):
-        img = self.image_np
-
-        annotations = Annotations.objects.filter(recording=self.recording)
-
-        threshold = 0.5
-        while True:
-            canvas = cv2.cvtColor(img, cv2.COLOR_RGB2GRAY)
-            canvas = canvas.astype(np.float32)
-
-            is_light = np.median(canvas) > 128.0
-            if is_light:
-                canvas = 255.0 - canvas
-
-            amplitude = canvas.max(axis=0)
-            amplitude -= amplitude.min()
-            amplitude /= amplitude.max()
-            amplitude[amplitude < threshold] = 0.0
-            amplitude[amplitude > 0] = 1.0
-            amplitude = amplitude.reshape(1, -1)
-
-            canvas -= canvas.min()
-            canvas /= canvas.max()
-            canvas *= 255.0
-            canvas *= amplitude
-            canvas = np.around(canvas).astype(np.uint8)
-
-            width = canvas.shape[1]
-            for annotation in annotations:
-                start = annotation.start_time / self.duration
-                stop = annotation.end_time / self.duration
-
-                start = int(np.around(start * width))
-                stop = int(np.around(stop * width))
-                canvas[:, start : stop + 1] = 255.0
-
-            mask = canvas.max(axis=0)
-            mask = scipy.signal.medfilt(mask, 3)
-            mask[0] = 0
-            mask[-1] = 0
-
-            starts = []
-            stops = []
-            for index in range(1, len(mask) - 1):
-                value_pre = mask[index - 1]
-                value = mask[index]
-                value_post = mask[index + 1]
-                if value != 0:
-                    if value_pre == 0:
-                        starts.append(index)
-                    if value_post == 0:
-                        stops.append(index)
-            assert len(starts) == len(stops)
-
-            starts = [val - 40 for val in starts]  # 10 ms buffer
-            stops = [val + 40 for val in stops]  # 10 ms buffer
-            ranges = list(zip(starts, stops))
-
-            while True:
-                found = False
-                merged = []
-                index = 0
-                while index < len(ranges) - 1:
-                    start1, stop1 = ranges[index]
-                    start2, stop2 = ranges[index + 1]
-
-                    start1 = min(max(start1, 0), len(mask))
-                    start2 = min(max(start2, 0), len(mask))
-                    stop1 = min(max(stop1, 0), len(mask))
-                    stop2 = min(max(stop2, 0), len(mask))
-
-                    if stop1 >= start2:
-                        found = True
-                        merged.append((start1, stop2))
-                        index += 2
-                    else:
-                        merged.append((start1, stop1))
-                        index += 1
-                    if index == len(ranges) - 1:
-                        merged.append((start2, stop2))
-                ranges = merged
-                if not found:
-                    for index in range(1, len(ranges)):
-                        start1, stop1 = ranges[index - 1]
-                        start2, stop2 = ranges[index]
-                        assert start1 < stop1
-                        assert start2 < stop2
-                        assert start1 < start2
-                        assert stop1 < stop2
-                        assert stop1 < start2
-                    break
-
-            segments = []
-            starts_ = []
-            stops_ = []
-            domain = img.shape[1]
-            widths = []
-            total_width = 0
-            for start, stop in ranges:
-                segment = img[:, start:stop]
-                segments.append(segment)
-
-                starts_.append(int(round(self.duration * (start / domain))))
-                stops_.append(int(round(self.duration * (stop / domain))))
-                widths.append(stop - start)
-                total_width += stop - start
-
-                # buffer = np.zeros((len(img), 20, 3), dtype=img.dtype)
-                # segments.append(buffer)
-            # segments = segments[:-1]
-
-            if len(segments) > 0:
-                break
-
-            threshold -= 0.05
-            if threshold < 0:
-                segments = None
-                break
-
-        if segments is None:
-            canvas = img.copy()
-        else:
-            canvas = np.hstack(segments)
-
-        canvas = Image.fromarray(canvas, 'RGB')
-
-        # canvas.save('temp.compressed.jpg')
-
-        buf = io.BytesIO()
-        canvas.save(buf, format='JPEG', quality=80)
-        buf.seek(0)
-        img_base64 = base64.b64encode(buf.getvalue()).decode('utf-8')
-
-        metadata = {
-            'starts': starts_,
-            'stops': stops_,
-            'widths': widths,
-            'length': total_width,
-        }
-        return canvas, img_base64, metadata
+        return spectrogram.pk
 
     @property
     def image_np(self):
@@ -395,77 +255,12 @@ class Spectrogram(TimeStampedModel, models.Model):
 
         return img_base64
 
-    def predict(self):
-        import json
-        import os
+    @property
+    def image_url(self):
+        return default_storage.url(self.image_file.name)
 
-        import onnx
-        import onnxruntime as ort
-        import tqdm
 
-        img, _, _ = self.compressed
-
-        relative = ('..',) * 4
-        asset_path = os.path.abspath(os.path.join(__file__, *relative, 'assets'))
-
-        onnx_filename = os.path.join(asset_path, 'model.mobilenet.onnx')
-        assert os.path.exists(onnx_filename)
-
-        session = ort.InferenceSession(
-            onnx_filename,
-            providers=[
-                (
-                    'CUDAExecutionProvider',
-                    {
-                        'cudnn_conv_use_max_workspace': '1',
-                        'device_id': 0,
-                        'cudnn_conv_algo_search': 'HEURISTIC',
-                    },
-                ),
-                'CPUExecutionProvider',
-            ],
-        )
-
-        img = np.array(img)
-
-        h, w, c = img.shape
-        ratio_y = 224 / h
-        ratio_x = ratio_y * 0.5
-        raw = cv2.resize(img, None, fx=ratio_x, fy=ratio_y, interpolation=cv2.INTER_LANCZOS4)
-
-        h, w, c = raw.shape
-        if w <= h:
-            canvas = np.zeros((h, h + 1, 3), dtype=raw.dtype)
-            canvas[:, :w, :] = raw
-            raw = canvas
-            h, w, c = raw.shape
-
-        inputs_ = []
-        for index in range(0, w - h, 100):
-            inputs_.append(raw[:, index : index + h, :])
-        inputs_.append(raw[:, -h:, :])
-        inputs_ = np.array(inputs_)
-
-        chunksize = 1
-        chunks = np.array_split(inputs_, np.arange(chunksize, len(inputs_), chunksize))
-        outputs = []
-        for chunk in tqdm.tqdm(chunks, desc='Inference'):
-            outputs_ = session.run(
-                None,
-                {'input': chunk},
-            )
-            outputs.append(outputs_[0])
-        outputs = np.vstack(outputs)
-        outputs = outputs.mean(axis=0)
-
-        model = onnx.load(onnx_filename)
-        mapping = json.loads(model.metadata_props[0].value)
-        labels = [mapping['forward'][str(index)] for index in range(len(mapping['forward']))]
-
-        prediction = np.argmax(outputs)
-        label = labels[prediction]
-        score = outputs[prediction]
-
-        confs = dict(zip(labels, outputs))
-
-        return label, score, confs
+@receiver(models.signals.pre_delete, sender=Spectrogram)
+def delete_content(sender, instance, **kwargs):
+    if instance.image_file:
+        instance.image_file.delete(save=False)
