+++ bats_ai/core/rest/__init__.py
@@ -1,10 +1,8 @@
 from rest_framework import routers
 
-from .image import ImageViewSet
 from .spectrogram import SpectrogramViewSet
 
-__all__ = ['ImageViewSet', 'SpectrogramViewSet']
+__all__ = ['SpectrogramViewSet']
 
 rest = routers.SimpleRouter()
-rest.register(r'images', ImageViewSet)
 rest.register(r'spectrograms', SpectrogramViewSet)
